requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==21.4.0
chardet==3.0.4
colorama==0.4.5
discord.py==1.7.3
pyperclip==1.8.2
google==3.0.0
idna==3.3
idna-ssl==1.1.0
multidict==4.7.6
typing-extensions==3.10.0.2
yarl==1.5.1
discord-components==2.1.2
python-dotenv==0.20.0
tk==0.1.0